Set verbose as the default behavior (#250)
Make Sherpa Slackbot messages verbose by default, to aid in debugging and observability.

This commit replaces the `verbose` option with a new option `concise`.  Using the concise 
option disables verbose messaging. It's opposites day!

Usage (verbose mode, which is now the default):
"What is AutoGPT and how does it compare with MetaGPT"

Usage (concise mode, to reduce verbose output):
"What is AutoGPT and how does it compare with MetaGPT --concise"