Fix Slackapp tests to run with `make test`
Make changes to fix broken tests and enable Slackapp tests to run
from `make test` with src/apps/slackapp as the working directory.

Give Slackapp a conftest.py of its own. This file is a copy of
Sherpa's conftest. Is there a clean way to reference it directly,
instead of copying it?

Change bolt_app to reference Sherpa's configuration file using
a relative file path.