refactor combinedID to be present out of the sherpa package (#311)
Prior to this change, Sherpa identifies individual users with the combination of two fields: team_id and user_id . These fields are both passed around between many functions. However these are Slack-specific team and user identities. We need to change this, because the core Sherpa library should not have a dependency on Slack. We want Sherpa to be able to work with many different user interfaces, all of which will have different authentication schemes.

Changes:

1. Change the Slack app so that when it passes user identity to any other part of Sherpa it combines team_id and user_id into a single string parameter. These strings should look like this: "slack" + slack_team_id + _ + slack_user_id.  The "slack" prefix will prevent collisions between user IDs if we add a new authentication mechanism in the future.  So we will encounter combined IDs like, slack_1234_999000111 (for team 1234 and user 999000111).

2. Change all other parts of Sherpa, including the UserUsageTracker, to expect a single string user_id parameter. Treat this ID value as opaque and assume that each ID is globally unique.

3. Remove team_id from all parts of Sherpa code, excepting the Slack app. 

When deploying we'll need to modify the data in the usage tracker database to account for this change, or just throw away the whitelist and rebuild it (since it is very small).

Also:

* Require user IDs in Whitelist table to be uniqe
* Rework tests to mock s3 client and reset database tables
* Change UserUsageTracker to permit passing in a Session and Engine, so
that we can use test-defined classes.
* Always cast self.limit_time_size_in_hours to float.
* Tidy up: run make format and make lint

Closes SHX-38
---------

Co-authored-by: Oshoma Momoh <oshoma@gmail.com>