Refactor tools.py and add LinkScraperTool
This commit refactors the `tools.py` file by organizing imports and removing unused imports. It also adds a new class `LinkScraperTool` which allows accessing the content of a link and extracting information from it. This tool can be used when there is a need to scrape data from a link.